# docs/swagger.yaml
basePath: /
definitions:
  dto.CreationEventDto:
    properties:
      deadlineDate:
        type: string
      description:
        type: string
      link:
        type: string
    type: object
  dto.UpdateEventDto:
    properties:
      deadlineDate:
        type: string
      description:
        type: string
      isCompleted:
        type: boolean
      link:
        type: string
    type: object
  er.ErrResponse:
    properties:
      error:
        type: string
      message:
        type: string
      status_code:
        type: integer
    type: object
  resp.Response:
    properties:
      body: {}
      status_code:
        type: integer
    type: object
host: localhost:8081
info:
  contact: {}
  description: API server for Mizar application
  title: Mizar API
  version: "1.0"
paths:
  /events:
    get:
      consumes:
      - application/json
      description: get all events
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/resp.Response'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/er.ErrResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/er.ErrResponse'
      summary: GetAllEvents
      tags:
      - events
    post:
      consumes:
      - application/json
      description: create new event
      parameters:
      - description: json with data about a new event
        in: body
        name: DataToCreate
        required: true
        schema:
          $ref: '#/definitions/dto.CreationEventDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/er.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/er.ErrResponse'
      summary: Create
      tags:
      - events
  /events/{eventId}:
    delete:
      consumes:
      - application/json
      description: delete event by id
      parameters:
      - description: Event id
        in: path
        name: eventId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/er.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/er.ErrResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/er.ErrResponse'
      summary: Delete
      tags:
      - events
    get:
      consumes:
      - application/json
      description: get event by id
      parameters:
      - description: Event id
        in: path
        name: eventId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/er.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/er.ErrResponse'
      summary: GetById
      tags:
      - events
    patch:
      consumes:
      - application/json
      description: update event by id
      parameters:
      - description: Event id
        in: path
        name: eventId
        required: true
        type: integer
      - description: json with data about a new event
        in: body
        name: DataToUpdate
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateEventDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/er.ErrResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/er.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/er.ErrResponse'
      summary: Update
      tags:
      - events
swagger: "2.0"